[GH-90] rename util to ops
