+++ impyute/contrib/describe.py
@@ -1,5 +1,5 @@
 """ impyute.contrib.describe """
-from impyute.util import find_null
+from impyute.ops import find_null
 
 def describe(data): # verbose=True):
     """ Print input/output multiple times
