+++ impyute/deletion/complete_case.py
@@ -1,7 +1,7 @@
 """ impyute.deletion.complete_case """
 import numpy as np
-from impyute.util import checks
-from impyute.util import preprocess
+from impyute.ops import checks
+from impyute.ops import preprocess
 
 @preprocess
 @checks
