+++ impyute/imputation/cs/em.py
@@ -1,8 +1,7 @@
-""" impyute.imputation.cs.em"""
 import numpy as np
-from impyute.util import find_null
-from impyute.util import preprocess
-from impyute.util import checks
+from impyute.ops import find_null
+from impyute.ops import preprocess
+from impyute.ops import checks
 
 @preprocess
 @checks
