+++ impyute/imputation/ts/locf.py
@@ -1,9 +1,8 @@
-""" impyute.imputation.ts.locf """
 import numpy as np
-from impyute.util import find_null
-from impyute.util import checks
-from impyute.util import preprocess
-from impyute.util.errors import BadInputError
+from impyute.ops import find_null
+from impyute.ops import checks
+from impyute.ops import preprocess
+from impyute.ops.errors import BadInputError
 @preprocess
 @checks
 def locf(data, axis=0):
