+++ impyute/ops/errors.py
@@ -1,4 +1,4 @@
-""" impyute.util.errors """
+""" Impyute specific error messages """
 
 class BadInputError(Exception):
     "Error thrown when input args don't match spec"
