+++ impyute/ops/util.py
