+++ test/imputation/cs/test_central_tendency.py
@@ -1,6 +1,6 @@
 """test_averagings.py"""
 import impyute as impy
-from impyute.util.testing import return_na_check
+from impyute.ops.testing import return_na_check
 
 SHAPE = (5, 5)
