+++ test/imputation/cs/test_em.py
@@ -1,6 +1,6 @@
 """test_em.py"""
 import impyute as impy
-from impyute.util.testing import return_na_check
+from impyute.ops.testing import return_na_check
 
 SHAPE = (5, 5)
