+++ test/imputation/ts/test_locf.py
@@ -1,8 +1,8 @@
 """test_locf.py"""
 import numpy as np
 import impyute as impy
-from impyute.util.testing import return_na_check
-from impyute.util.errors import BadInputError
+from impyute.ops.testing import return_na_check
+from impyute.ops.errors import BadInputError
 
 SHAPE = (5, 5)
