+++ test/imputation/ts/test_moving_window.py
@@ -2,7 +2,7 @@
 import pytest
 import numpy as np
 import impyute as impy
-from impyute.util.testing import return_na_check
+from impyute.ops.testing import return_na_check
 #pylint:disable=missing-docstring, redefined-outer-name
